backend/pyproject.toml
# backend/pyproject.toml

[tool.poetry]
name = "modeling-platform-backend"
version = "1.0.0"
description = "Enterprise Modeling Platform Backend API"
authors = ["Your Organization <dev@yourorg.com>"]
license = "MIT"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"

# Web Framework
fastapi = "^0.115.0"
uvicorn = {extras = ["standard"], version = "^0.32.0"}
python-multipart = "^0.0.18"
pydantic = "^2.10.0"
pydantic-settings = "^2.6.0"
email-validator = "^2.2.0"  # ADDED - Required for pydantic EmailStr validation

# Database - PostgreSQL
sqlalchemy = "^2.0.36"
alembic = "^1.14.0"
psycopg2-binary = "^2.9.10"
asyncpg = "^0.30.0"

# Database - Redis (Updated to latest)
redis = "^7.1.0"

# Database - FalkorDB (Updated to latest available version)
falkordb = "^1.2.2"

# Authentication & Security
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
bcrypt = "5.0.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-dateutil = "^2.9.0"


# WebSocket
websockets = "^14.1"
python-socketio = "^5.12.0"

# Parsing & Export
sqlglot = "^25.30.0"
lxml = "^5.3.0"
jinja2 = "^3.1.4"

# Utilities
pyyaml = "^6.0.2"
python-dotenv = "^1.0.1"
orjson = "^3.10.12"
httpx = "^0.28.1"
tenacity = "^9.0.0"

# Collaboration & CRDT
y-py = "^0.6.2"

# Validation
jsonschema = "^4.23.0"

# Monitoring & Logging
structlog = "^24.4.0"
python-json-logger = "^3.2.1"

[tool.poetry.group.dev.dependencies]
# Testing
pytest = "^8.3.4"
pytest-asyncio = "^0.24.0"
pytest-cov = "^6.0.0"
pytest-mock = "^3.14.0"
faker = "^33.1.0"

# Code Quality
black = "^24.10.0"
ruff = "^0.8.4"
mypy = "^1.13.0"
isort = "^5.13.2"
pre-commit = "^4.0.1"

# Type Stubs
types-redis = "^4.6.0.20241004"
types-pyyaml = "^6.0.12.20240917"
types-python-jose = "^3.3.4.20240106"

# Development Tools
ipython = "^8.31.0"
ipdb = "^0.13.13"

[tool.poetry.group.docs.dependencies]
mkdocs = "^1.6.1"
mkdocs-material = "^9.5.49"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
  | migrations
)/
'''

[tool.ruff]
line-length = 100
target-version = "py311"
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "C",  # flake8-comprehensions
    "B",  # flake8-bugbear
    "UP", # pyupgrade
]
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_calls = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false
disallow_untyped_calls = false

[tool.pytest.ini_options]
minversion = "8.0"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--strict-markers",
    "--cov=app",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--cov-report=xml",
]
asyncio_mode = "auto"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
]

[tool.coverage.run]
source = ["app"]
omit = [
    "*/tests/*",
    "*/migrations/*",
    "*/__init__.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@abstractmethod",
]

[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true